requirements.txt
azure-storage-blob 
azure-identity 
azure-search-documents==11.6.0b4 
azure-ai-inference==1.0.0b6 
azure-ai-documentintelligence==1.0.0 
python-dotenv 
pdf2image 
numpy
langchain 
tenacity 
nest_asyncio
